pyproject.toml
# Build System
[build-system]
  requires = ["uv_build"]
  build-backend = "uv_build"

# Project Metadata
[project]
  name = "lnkr-api"
  version = "0.2.1"
  description = "Link manager REST API."
  authors = [{ name = "Daniel Mizsak", email = "info@pythonvilag.hu" }]
  requires-python = ">=3.14"
  dependencies = [
    "PyJWT==2.11.0",
    "SQLAlchemy==2.0.48",
    "fastapi[standard]==0.135.1",
    "psycopg[binary]==3.3.3",
    "pydantic-settings==2.13.1",
    "redis==7.3.0",
    "sentry-sdk==2.54.0",
  ]

  [project.urls]
    Homepage = "https://api.lnkr.by"
    Repository = "https://github.com/daniel-mizsak/lnkr-api"

[dependency-groups]
  dev = [
    "beautifulsoup4",
    "coverage",
    "fakeredis",
    "pre-commit",
    "pytest",
    "ruff",
    "ty",
  ]
  deployment = ["ansible"]

# Tools
# Uv
[tool.uv.build-backend]
  module-root = "src"
  module-name = "lnkr"
